+++ cppcache/src/PdxLocalWriter.hpp
@@ -112,14 +112,14 @@ class PdxLocalWriter : public PdxWriter {
 
   template <typename mType>
   void writeArrayObject(std::vector<mType> array) {
-    m_dataOutput->writeArrayLen(array.size());
+    m_dataOutput->writeArrayLen(static_cast<int32_t>(array.size()));
     for (auto&& obj : array) {
       writeObject(obj);
     }
   }
 
   virtual PdxWriter& writeChar(const std::string& fieldName,
-                                               char16_t value) override;
+                               char16_t value) override;
 
   virtual PdxWriter& writeBoolean(const std::string& fieldName,
                                   bool value) override;
@@ -145,44 +145,36 @@ class PdxLocalWriter : public PdxWriter {
   virtual PdxWriter& writeDate(const std::string& fieldName,
                                std::shared_ptr<CacheableDate> date) override;
 
-  virtual PdxWriter& writeString(
-      const std::string& fieldName, const std::string& value) override;
+  virtual PdxWriter& writeString(const std::string& fieldName,
+                                 const std::string& value) override;
 
-  virtual PdxWriter& writeObject(
-      const std::string& fieldName,
-      std::shared_ptr<Serializable> value) override;
+  virtual PdxWriter& writeObject(const std::string& fieldName,
+                                 std::shared_ptr<Serializable> value) override;
 
-  virtual PdxWriter& writeBooleanArray(
-      const std::string& fieldName,
-      const std::vector<bool>& array) override;
+  virtual PdxWriter& writeBooleanArray(const std::string& fieldName,
+                                       const std::vector<bool>& array) override;
 
   virtual PdxWriter& writeCharArray(
       const std::string& fieldName,
       const std::vector<char16_t>& array) override;
 
-  virtual PdxWriter& writeByteArray(
-      const std::string& fieldName,
-      const std::vector<int8_t>& array) override;
+  virtual PdxWriter& writeByteArray(const std::string& fieldName,
+                                    const std::vector<int8_t>& array) override;
 
   virtual PdxWriter& writeShortArray(
-      const std::string& fieldName,
-      const std::vector<int16_t>& array) override;
+      const std::string& fieldName, const std::vector<int16_t>& array) override;
 
-  virtual PdxWriter& writeIntArray(
-      const std::string& fieldName,
-      const std::vector<int32_t>& array) override;
+  virtual PdxWriter& writeIntArray(const std::string& fieldName,
+                                   const std::vector<int32_t>& array) override;
 
-  virtual PdxWriter& writeLongArray(
-      const std::string& fieldName,
-      const std::vector<int64_t>& array) override;
+  virtual PdxWriter& writeLongArray(const std::string& fieldName,
+                                    const std::vector<int64_t>& array) override;
 
-  virtual PdxWriter& writeFloatArray(
-      const std::string& fieldName,
-      const std::vector<float>& array) override;
+  virtual PdxWriter& writeFloatArray(const std::string& fieldName,
+                                     const std::vector<float>& array) override;
 
   virtual PdxWriter& writeDoubleArray(
-      const std::string& fieldName,
-      const std::vector<double>& array) override;
+      const std::string& fieldName, const std::vector<double>& array) override;
 
   virtual PdxWriter& writeStringArray(
       const std::string& fieldName,
@@ -192,9 +184,10 @@ class PdxLocalWriter : public PdxWriter {
       const std::string& fieldName,
       std::shared_ptr<CacheableObjectArray> array) override;
 
-  virtual PdxWriter& writeArrayOfByteArrays(
-      const std::string& fieldName, int8_t* const* const array, int arrayLength,
-      const int* elementLength) override;
+  virtual PdxWriter& writeArrayOfByteArrays(const std::string& fieldName,
+                                            int8_t* const* const array,
+                                            int arrayLength,
+                                            const int* elementLength) override;
 
   virtual PdxWriter& markIdentityField(const std::string& fieldName) override;
